Cache Unset

Unsets a value from the cache.

cache.unset(cache, key)

Arguments

cache
cache

key
value key

Examples

# by default a memory cache will be created
myCache <- cache()

# the cache storage type can be specified, but only 'memory' is currently supported
alsoMyCache <- cache(storage = "memory")

# values can be stored in the cache 
cache.set(myCache, "name", "Pete Smith") 

#> $memory_cache
#> <environment: 0x000000001486fe58>
#> 
#> $storage
#> [1] "memory"
#> attr(,"class")
#> [1] "memory"
#> 
#> attr(,"class")
#> [1] "cache"

cache.set(myCache, "age", "72")

#> $memory_cache
#> <environment: 0x000000001486fe58>
#> 
#> $storage
#> [1] "memory"
#> attr(,"class")
#> [1] "memory"
#> 
#> attr(,"class")
#> [1] "cache"

# and retrieved
print(paste(cache.get(myCache, "name"), "is", cache.get(myCache, "age"), "years old."))

#> [1] "Pete Smith is 72 years old."

# setting a value with key that has already been used overwrites the currently stored value
cache.set(myCache, "name", "Sam Smith")

#> $memory_cache
#> <environment: 0x000000001486fe58>
#> 
#> $storage
#> [1] "memory"
#> attr(,"class")
#> [1] "memory"
#> 
#> attr(,"class")
#> [1] "cache"

print(paste(cache.get(myCache, "name"), "is", cache.get(myCache, "age"), "years old."))

#> [1] "Sam Smith is 72 years old."

# set keys can be detected
if (cache.has(myCache, "name")) print("value for name key is set")

#> [1] "value for name key is set"

# keys can be unset
cache.unset(myCache, "name")
if (!cache.has(myCache, "name")) print("value for name key is not set")

#> [1] "value for name key is not set"

# all values in the cache can be cleared
cache.clear(myCache)